+++ packages/web/src/lib/posthogEvents.ts
@@ -216,6 +216,10 @@ export type PosthogEventMap = {
     //////////////////////////////////////////////////////////////////
     wa_login_with_github: {},
     wa_login_with_google: {},
+    wa_login_with_gitlab: {},
+    wa_login_with_okta: {},
+    wa_login_with_keycloak: {},
+    wa_login_with_microsoft_entra_id: {},
     wa_login_with_magic_link: {},
     wa_login_with_credentials: {},
     //////////////////////////////////////////////////////////////////