+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fixed issue where new oauth providers weren't being display in the login page
+
 ## [4.0.1] - 2025-05-28
 
 ### Fixed